Refactor PositionReader
